+++ internal/test-koin/src/commonMain/kotlin/com/github/panpf/sketch/test/utils/Koins.kt
@@ -9,7 +9,7 @@ object Koins {
     init {
         startKoin {
             modules(module {
-                single { Sketch.Builder(getTestContext()).build() }
+                single { Sketch(getTestContext()) }
             })
         }
     }