+++ sketch-singleton/src/commonTest/kotlin/com/github/panpf/sketch/singleton/common/test/SingletonSketchTest.kt
@@ -16,9 +16,9 @@ class SingletonSketchTest {
     fun test() {
         val context = getTestContext()
 
-        val mySketch1 = Sketch.Builder(context).build()
-        val mySketch2 = Sketch.Builder(context).build()
-        val mySketch3 = Sketch.Builder(context).build()
+        val mySketch1 = Sketch(context)
+        val mySketch2 = Sketch(context)
+        val mySketch3 = Sketch(context)
         assertFalse(mySketch1.isShutdown)
         assertFalse(mySketch2.isShutdown)
         assertFalse(mySketch3.isShutdown)